+++ agent/library/src/nativeMain/kotlin/org/matilda/Main.kt
@@ -3,11 +3,16 @@ package org.matilda
 import kotlinx.cinterop.ExperimentalForeignApi
 import org.matilda.logger.StderrLogger
 import org.matilda.utils.Fd
+import org.matilda.utils.getTracerId
 import org.matilda.utils.sink
 import org.matilda.utils.source
 import platform.posix.*
 
-fun main() {
+fun main(args: Array<String>) {
+    if (args.contains("--wait-for-debugger")) {
+        waitForDebugger()
+    }
+
     val logger = StderrLogger()
     try {
         pingToLoader()
@@ -17,6 +22,12 @@ fun main() {
         // but we don't want to write there because we use it
         logger.log("Error running Matilda agent", e)
         exit(-1)
+    }
+}
+
+fun waitForDebugger() {
+    while (getTracerId() == 0) {
+        sleep(1u)
     }
 }